requirements.txt
PyQt5==5.14.1
PyVISA-py==0.3.1
PyVISA==1.10.1
numpy==1.18.2